+++ app/lib/features/action/presentation/pages/action_page.dart
@@ -9,13 +9,11 @@ import 'package:app/features/action/presentation/bloc/action_bloc.dart';
 import 'package:app/features/action/presentation/bloc/action_event.dart';
 import 'package:app/features/action/presentation/bloc/action_state.dart';
 import 'package:app/features/actions/domain/action_type.dart';
-import 'package:app/features/aids/core/domain/aid.dart';
 import 'package:app/features/aids/core/domain/aid_summary.dart';
-import 'package:app/features/aids/core/presentation/widgets/aid_card.dart';
+import 'package:app/features/aids/core/presentation/widgets/aid_summary_card.dart';
 import 'package:app/features/car_simulator/presentation/widgets/car_simulator_widget.dart';
 import 'package:app/features/services/lvao/presentation/widgets/lvao_horizontal_list.dart';
 import 'package:app/features/services/recipes/action/presentation/widgets/recipe_horizontal_list.dart';
-import 'package:app/features/theme/core/domain/theme_type.dart';
 import 'package:app/l10n/l10n.dart';
 import 'package:dsfr/dsfr.dart';
 import 'package:flutter/material.dart';
