+++ app/lib/features/aids/item/presentation/bloc/aid_bloc.dart
@@ -5,8 +5,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 
 class AidBloc extends Bloc<AidEvent, AidState> {
   AidBloc({required final AidsRepository repository}) : super(const AidStateLoading()) {
-    on<AidSelected>((final event, final emit) => emit(AidStateSuccess(event.value)));
-
     on<AidSelectedById>((final event, final emit) async {
       final aidResult = await repository.fetchById(event.aidId);
